implement composing objects
Object composition is now super cheap since we use a content addressable
store and we allow blobs to be defined in terms of multiple chunks.

We may want to support composition down at the meta layer so we can do
this in a single transaction, but for now this will do.

Also fixed a typo where all objects were being named after their bucket,
oops!